requirements.txt
streamlit==1.40.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
google-generativeai==0.8.0
python-dotenv==1.0.0
openpyxl==3.1.2
scikit-learn==1.4.0
scipy==1.12.0
